requirements.txt
async-generator==1.10
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
outcome==1.1.0
pyfuse3==3.2.0
python-dotenv==0.15.0
requests==2.25.1
sniffio==1.2.0
sortedcontainers==2.3.0
trio==0.18.0
urllib3==1.26.3
